source/crypto_api_function/requirements.txt
google-cloud-secret-manager
requests
google-cloud-pubsub
